Samples: fix some broken GL3+ shaders

+++ Samples/Media/materials/programs/GLSL400/RasterizationOrderFp.glsl
@@ -1,5 +1,7 @@
 #version 400
 
+#extension GL_ARB_shader_atomic_counters : enable
+
 layout(binding=0, offset=0) uniform atomic_uint ac;
 
 out vec4 fragColour;
@@ -13,5 +15,5 @@ void main(void)
     //     ((counter & (mask <<  8)) % 255) / 255.f,
     //     ((counter & (mask << 16)) % 255) / 5.f,
     //     0.5);
-    fragColour = vec4(counter/4E9,0,0,0);
+    fragColour = vec4(counter/uint(4E9),0,0,0);
 }
